+++ lib/src/main/java/com/smileidentity/compose/ProcessingScreen.kt
@@ -260,7 +260,9 @@ private fun PreviewProcessingSuccessScreen() {
             icon = painterResource(R.drawable.si_processing_success),
             title = stringResource(R.string.si_smart_selfie_processing_success_title),
             subtitle = stringResource(R.string.si_smart_selfie_processing_success_subtitle),
-            continueButtonText = stringResource(R.string.si_smart_selfie_processing_continue_button),
+            continueButtonText = stringResource(
+                R.string.si_smart_selfie_processing_continue_button,
+            ),
             onContinue = {},
         )
     }