+++ lib/src/main/java/com/smileidentity/viewmodel/SelfieViewModel.kt
@@ -34,14 +34,19 @@ import com.smileidentity.results.SmileIDCallback
 import com.smileidentity.results.SmileIDResult
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.debounce
+import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import timber.log.Timber
 import java.io.File
 import kotlin.math.absoluteValue
+import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.Duration.Companion.seconds
 
+private val UI_DEBOUNCE_DURATION = 250.milliseconds
 private const val INTRA_IMAGE_MIN_DELAY_MS = 350
 private const val NUM_LIVENESS_IMAGES = 7
 private const val TOTAL_STEPS = NUM_LIVENESS_IMAGES + 1 // 7 B&W Liveness + 1 Color Selfie
@@ -65,6 +70,7 @@ enum class Directive(@StringRes val displayText: Int) {
     InitialInstruction(R.string.si_smart_selfie_instructions),
     Capturing(R.string.si_smart_selfie_directive_capturing),
     EnsureFaceInFrame(R.string.si_smart_selfie_directive_unable_to_detect_face),
+    EnsureOnlyOneFace(R.string.si_smart_selfie_directive_multiple_faces),
     MoveCloser(R.string.si_smart_selfie_directive_face_too_far),
     MoveAway(R.string.si_smart_selfie_directive_face_too_close),
     Smile(R.string.si_smart_selfie_directive_smile),
@@ -76,7 +82,13 @@ class SelfieViewModel(
     private val jobId: String,
 ) : ViewModel() {
     private val _uiState = MutableStateFlow(SelfieUiState())
-    val uiState = _uiState.asStateFlow()
+
+    // Debounce to avoid spamming Directive updates so that they can be read by the user
+    val uiState = _uiState.asStateFlow().debounce(UI_DEBOUNCE_DURATION).stateIn(
+        viewModelScope,
+        SharingStarted.WhileSubscribed(),
+        SelfieUiState(),
+    )
     var result: SmileIDResult<SmartSelfieResult>? = null
 
     private val livenessFiles = mutableListOf<File>()
@@ -127,21 +139,26 @@ class SelfieViewModel(
                 return@addOnSuccessListener
             }
 
-            // Pick the largest face
-            val largestFace = faces.maxBy { it.boundingBox.area }
-            val faceFillRatio = (largestFace.boundingBox.area / inputImage.area.toFloat())
+            // Ensure only 1 face is in frame
+            if (faces.size > 1) {
+                _uiState.update { it.copy(currentDirective = Directive.EnsureOnlyOneFace) }
+                return@addOnSuccessListener
+            }
+
+            val face = faces.first()
 
             // Check that the corners of the face bounding box are within the inputImage
-            val faceCornersInImage = largestFace.boundingBox.left >= 0 &&
-                largestFace.boundingBox.right <= inputImage.width &&
-                largestFace.boundingBox.top >= 0 &&
-                largestFace.boundingBox.bottom <= inputImage.height
+            val faceCornersInImage = face.boundingBox.left >= 0 &&
+                face.boundingBox.right <= inputImage.width &&
+                face.boundingBox.top >= 0 &&
+                face.boundingBox.bottom <= inputImage.height
             if (!faceCornersInImage) {
                 _uiState.update { it.copy(currentDirective = Directive.EnsureFaceInFrame) }
                 return@addOnSuccessListener
             }
 
             // Check that the face is close enough to the camera
+            val faceFillRatio = (face.boundingBox.area / inputImage.area.toFloat())
             if (faceFillRatio < MIN_FACE_AREA_THRESHOLD) {
                 _uiState.update { it.copy(currentDirective = Directive.MoveCloser) }
                 return@addOnSuccessListener
@@ -154,7 +171,7 @@ class SelfieViewModel(
             }
 
             // Ask the user to start smiling partway through liveness images
-            val isSmiling = (largestFace.smilingProbability ?: 0f) > SMILE_THRESHOLD
+            val isSmiling = (face.smilingProbability ?: 0f) > SMILE_THRESHOLD
             if (livenessFiles.size > NUM_LIVENESS_IMAGES / 2 && !isSmiling) {
                 _uiState.update { it.copy(currentDirective = Directive.Smile) }
                 return@addOnSuccessListener
@@ -164,13 +181,13 @@ class SelfieViewModel(
 
             // Perform the rotation checks *after* changing directive to Capturing -- we don't want
             // to explicitly tell the user to move their head
-            if (!hasFaceRotatedEnough(largestFace)) {
+            if (!hasFaceRotatedEnough(face)) {
                 Timber.v("Not enough face rotation between captures. Waiting...")
                 return@addOnSuccessListener
             }
-            previousHeadRotationX = largestFace.headEulerAngleX
-            previousHeadRotationY = largestFace.headEulerAngleY
-            previousHeadRotationZ = largestFace.headEulerAngleZ
+            previousHeadRotationX = face.headEulerAngleX
+            previousHeadRotationY = face.headEulerAngleY
+            previousHeadRotationZ = face.headEulerAngleZ
 
             // TODO: CameraX 1.3.0-alpha04 adds built-in API to convert ImageProxy to Bitmap.
             //  Incorporate once stable