+++ sample/src/main/java/com/smileidentity/sample/compose/MainScreen.kt
@@ -46,6 +46,9 @@ import com.smileidentity.compose.DocumentVerification
 import com.smileidentity.compose.SmartSelfieAuthenticationScreen
 import com.smileidentity.compose.SmartSelfieRegistrationScreen
 import com.smileidentity.models.Document
+import com.smileidentity.compose.SmartSelfieEnrollmentScreen
+import com.smileidentity.models.Document
+import com.smileidentity.randomJobId
 import com.smileidentity.randomUserId
 import com.smileidentity.results.SmileIDResult
 import com.smileidentity.sample.BottomNavigationScreen
@@ -115,7 +118,9 @@ fun MainScreen() {
                                     if (isProduction) {
                                         Icon(
                                             imageVector = Icons.Filled.Warning,
-                                            contentDescription = stringResource(R.string.production),
+                                            contentDescription = stringResource(
+                                                R.string.production,
+                                            ),
                                         )
                                     }
                                 },
@@ -192,17 +197,17 @@ fun MainScreen() {
                             currentScreenTitle = BottomNavigationScreen.AboutUs.label
                             AboutUsScreen()
                         }
-                        composable(ProductScreen.SmartSelfieRegistration.route) {
+                        composable(ProductScreen.SmartSelfieEnrollment.route) {
                             bottomNavSelection = BottomNavigationScreen.Home
-                            currentScreenTitle = ProductScreen.SmartSelfieRegistration.label
+                            currentScreenTitle = ProductScreen.SmartSelfieEnrollment.label
                             val userId = rememberSaveable { randomUserId() }
-                            SmileID.SmartSelfieRegistrationScreen(
+                            SmileID.SmartSelfieEnrollmentScreen(
                                 userId = userId,
                                 allowAgentMode = true,
                             ) { result ->
                                 if (result is SmileIDResult.Success) {
                                     val resultData = result.data
-                                    val message = StringBuilder("SmartSelfie Registration ")
+                                    val message = StringBuilder("SmartSelfie Enrollment ")
                                     if (resultData.jobStatusResponse.jobComplete) {
                                         if (resultData.jobStatusResponse.jobSuccess) {
                                             message.append("completed successfully. ")
@@ -221,7 +226,7 @@ fun MainScreen() {
                                     )
                                 } else if (result is SmileIDResult.Error) {
                                     val th = result.throwable
-                                    val message = "SmartSelfie Registration error: ${th.message}"
+                                    val message = "SmartSelfie Enrollment error: ${th.message}"
                                     Timber.e(th, message)
                                     snackbarHostState.showSnackbar(coroutineScope, message)
                                 }
@@ -262,7 +267,7 @@ fun MainScreen() {
                                         enabled = userId.isNotBlank(),
                                         onClick = {
                                             navController.navigate(
-                                                "${ProductScreen.SmartSelfieAuthentication.route}/$userId",
+                                                "${ProductScreen.SmartSelfieAuthentication.route}/$userId", // ktlint-disable max-line-length
                                             ) { popUpTo(BottomNavigationScreen.Home.route) }
                                         },
                                     ) { Text(stringResource(R.string.cont)) }
@@ -325,6 +330,7 @@ fun MainScreen() {
                         composable(ProductScreen.DocumentVerification.route) {
                             bottomNavSelection = BottomNavigationScreen.Home
                             currentScreenTitle = ProductScreen.DocumentVerification.label
+
                             SmileID.DocumentVerification(
                                 allowGalleryUpload = false,
                                 captureBothSides = true,
@@ -335,13 +341,63 @@ fun MainScreen() {
                                     val message = "Document Verification success"
                                     Timber.d("$message: $resultData")
                                     snackbarHostState.showSnackbar(coroutineScope, message)
+
+                            // TODO: fetch this from
+                            //  https://api.smileidentity.com/v1/products_config once it supports
+                            //  signature auth (there should be a story for this for Platform)
+                            DocumentVerificationIdTypeSelector(docVTestData) { country, idType ->
+                                Timber.v("Selected country: $country, idType: $idType")
+                                navController.navigate(
+                                    "${ProductScreen.DocumentVerification.route}/$country/$idType",
+                                ) { popUpTo(ProductScreen.DocumentVerification.route) }
+                            }
+                        }
+                        composable(
+                            ProductScreen.DocumentVerification.route + "/{countryCode}/{idType}",
+                        ) {
+                            val userId = rememberSaveable { randomUserId() }
+                            val jobId = rememberSaveable { randomJobId() }
+                            val documentType = Document(
+                                it.arguments?.getString("countryCode")!!,
+                                it.arguments?.getString("idType")!!,
+                            )
+                            SmileID.DocumentVerification(
+                                userId = userId,
+                                jobId = jobId,
+                                enforcedIdType = documentType,
+                            ) { result ->
+                                if (result is SmileIDResult.Success) {
+                                    val resultData = result.data
+                                    val message = StringBuilder("Document Verification ")
+                                    if (resultData.jobStatusResponse.jobComplete) {
+                                        if (resultData.jobStatusResponse.jobSuccess) {
+                                            message.append("completed successfully.")
+                                        } else {
+                                            message.append("completed unsuccessfully.")
+                                        }
+                                    } else {
+                                        message.append("still pending.")
+                                    }
+                                    Timber.d("$message: $result")
+                                    snackbarHostState.showSnackbar(
+                                        coroutineScope,
+                                        message.toString(),
+                                    )
+
                                 } else if (result is SmileIDResult.Error) {
                                     val th = result.throwable
                                     val message = "Document Verification error: ${th.message}"
                                     Timber.e(th, message)
                                     snackbarHostState.showSnackbar(coroutineScope, message)
                                 }
+
                                 navController.popBackStack()
+
+                                navController.popBackStack(
+                                    route = BottomNavigationScreen.Home.route,
+                                    inclusive = false,
+                                )
+
                             }
                         }
                     }