+++ CHANGELOG.md
@@ -3,7 +3,40 @@
 ## Unreleased
 
 ### Features
-- Initial v10 release 🎉
+- Breaking: renamed SmartSelfie Registration to SmartSelfie Enrollment
+- Compile against API level 34
+- Minor change in default colors to add contrast
+- Debounce Selfie Capture directive changes to allow user time to read the directions
+- Tweak selfie progress indicator animation
+- Set Release property on Sentry for release tracking
+
+### Bug Fixes
+- Fix crash on network retries
+
+### Dependencies
+- Bump AndroidX Activity to 1.7.2
+- Bump Gradle to 8.0.2
+
+## 10.0.0-beta02
+
+### Features
+- Add Biometric KYC data models
+- Initial Document Verification screens
+
+### Bug Fixes
+- Only allow a single face to be in frame for a SmartSelfie™ capture
+
+### Dependencies
+- Bump Sentry to 6.21.0
+
+## 10.0.0-beta01
+
+### Features
+- Initial release 🎉
+- SmartSelfie™ Authentication and Registration
+- Enhanced KYC
+- Theming
+- Networking
 
 ### Bug Fixes
 N/A